requirements.txt
aiomysql
vkbottle
loguru
tortoise-orm
emoji
django-environ
